Added new files

+++ utils/mymath.h
@@ -66,7 +66,7 @@ const T CMath<T>::MAXREAL=std::numeric_limits<T>::max();
 template<class T>
 const T CMath<T>::MINREAL=std::numeric_limits<T>::min();
 
-typedef double Real;
+typedef float Real;
